ajustes en lectura de corriente

+++ Truma_EBL_Bios_V100/Truma_EBL_Bios_V100/Core/Inc/LeisureBattery1.h
@@ -30,7 +30,7 @@ public:
 		void processInfoFrame(uint8_t* frame);
 
 		int state;
-		int mAmps;
+		float mAmps;
 		int timeRemaining;
 		int soc;
 		int volts;
